backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
pydantic-settings>=2.2
python-dotenv>=1.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
joblib>=1.3
groq>=0.4
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
sqlalchemy>=2.0
aiosqlite>=0.19
twilio>=8.0
python-multipart>=0.0.9
